@@ -1,10 +1,21 @@
 {
   "name": "@sap-devx/inquirer-gui-radio-plugin",
-  "version": "0.0.1",
+  "version": "3.0.0",
   "description": "Sample radio plugin for inquirer-gui",
   "main": "dist/radioPlugin.umd.js",
+  "module": "./dist/radioPlugin.es.js",
+  "exports": {
+    ".": {
+      "import": "./dist/radioPlugin.es.js",
+      "require": "./dist/radioPlugin.umd.js"
+    },
+    "./dist/radioPlugin.css": {
+      "import": "./dist/radioPlugin.css",
+      "require": "./dist/radioPlugin.css"
+    }
+  },
   "scripts": {
-    "build": "vue-cli-service build --target lib --name radioPlugin src/index.js",
+    "build": "vite build",
     "prepublishOnly": "npm run build"
   },
   "repository": {
@@ -14,9 +25,10 @@
   "author": "Vered Tagor",
   "license": "Apache-2.0",
   "dependencies": {
-    "vue": "^2.6.10"
+    "vue": "^3.3.4"
   },
   "devDependencies": {
-    "@vue/cli-service": "^3.11.0"
+    "vite": "^4.3.9",
+    "@vitejs/plugin-vue": "^4.2.3"
   }
 }
